==> code/sirds_hybrid/03_extracting_time_series.ipynb <==
# Extracting the death forecasting time series

import pandas as pd

from code.sirds.sirds_model_visualization import get_sirds, get_sirds_extras
import datetime
import numpy as np

## Reading data

df_results = pd.read_csv('../sirds/data/estimation_results.csv')
df_results['max_date_to_fit'] = pd.to_datetime(df_results['max_date_to_fit']) 

df_results = df_results.sort_values(['max_date_to_fit','municipality'])

FORECASTING_HORIZON = 84

df_estimated_parameters = pd.read_csv('data/df_estimated_parameters.csv')
df_estimated_parameters['max_date_to_fit'] = pd.to_datetime(df_estimated_parameters['max_date_to_fit'])

## Executing results

df_results

df_deaths_forecasted = pd.DataFrame()
df_simulations = pd.DataFrame()
delta = datetime.timedelta(days=1)

for max_date_to_fit in df_results['max_date_to_fit'].unique()[:]:
    df_results_max_data_to_fit = df_results.loc[df_results['max_date_to_fit'] == max_date_to_fit]
    date_list_forecasting = [pd.to_datetime(max_date_to_fit) + (delta * (i + 1)) for i in range(FORECASTING_HORIZON)]    
    for id_municipality in df_results_max_data_to_fit.municipality_id.unique():        
        df_results_municipality = df_results_max_data_to_fit.loc[df_results_max_data_to_fit.municipality_id == id_municipality]
        municipality = df_results_municipality.municipality.iloc[0]
    
        for index, result in df_results_municipality.iterrows():
            result = result.dropna()
            
            estimation = result.estimation
                        
            estimated_parameters = df_estimated_parameters.loc[(df_estimated_parameters['municipality_id']==id_municipality) & (df_estimated_parameters['max_date_to_fit']==max_date_to_fit) & (df_estimated_parameters['estimation']==estimation)].iloc[0]
            estimate_fatality_in_forecast_period = estimated_parameters['estimated_fatality']
            estimated_contact_rate_in_forecast_period = estimated_parameters['estimated_contact_rate']
            if np.isnan(estimated_contact_rate_in_forecast_period):
                estimated_contact_rate_in_forecast_period = None
                estimated_days_between_infections_in_forecast_period = None
            else:
                estimated_days_between_infections_in_forecast_period = 1 / estimated_contact_rate_in_forecast_period
            
            if np.isnan(estimate_fatality_in_forecast_period):
                estimate_fatality_in_forecast_period = None
            
            date_first_case = pd.to_datetime(result.date_first_case)
            days_to_recovery = result.days_to_recovery
            period_in_days_to_fit = int(result.period_in_days)
            period_in_days_general = period_in_days_to_fit + FORECASTING_HORIZON
    
            y = get_sirds(result, forecast_horizon=FORECASTING_HORIZON, contact_rate_in_forecast_period=estimated_contact_rate_in_forecast_period, fatality_in_forecast_period=estimate_fatality_in_forecast_period)
            S, I, R, D, I_accumulated = y
            N = S[0] + I[0] + R[0] + D[0]
    
            D_new_deaths, reproduction_number_sird, I_new_cases, epidemic_periods_with_fast_transition_fuzzy_variable, epidemic_periods_with_slow_transition_fuzzy_variable, days_between_infections_values, case_fatality_probability_values, loss_immunity_in_days_values, estimated_days_between_infections, estimated_case_fatality_probability, estimated_loss_immunity_in_days = get_sirds_extras(result, S, D, I_accumulated, forecast_horizon=FORECASTING_HORIZON, estimated_days_between_infections_in_forecast_period=estimated_days_between_infections_in_forecast_period, estimated_fatality_in_forecast_period=estimate_fatality_in_forecast_period)
                            
            forecast = D_new_deaths[-FORECASTING_HORIZON:]
            df_forecasting_municipality = pd.DataFrame()
            df_forecasting_municipality['forecast'] = forecast
            df_forecasting_municipality['municipality'] = municipality
            df_forecasting_municipality['municipality_id'] = id_municipality
            df_forecasting_municipality['max_date_to_fit'] = max_date_to_fit
            df_forecasting_municipality['estimation'] = estimation            
            df_forecasting_municipality['date'] = date_list_forecasting
            df_deaths_forecasted = pd.concat([df_deaths_forecasted, df_forecasting_municipality])
            
            df_simulation = pd.DataFrame()
            df_simulation['S'] = S[1:]
            df_simulation['I'] = I[1:]
            df_simulation['R'] = R[1:]
            df_simulation['D_new_deaths'] = D_new_deaths
            df_simulation['effective_reproduction_number'] = reproduction_number_sird[1:]
            R0 = days_to_recovery/estimated_days_between_infections
            df_simulation['basic_reproduction_number'] = R0[1:]
            df_simulation['infection_fatality_rate'] = estimated_case_fatality_probability[1:]
            date_list_simulation = [pd.to_datetime(date_first_case) + (delta * (i + 1)) for i in range(period_in_days_general-1)]
            df_simulation['date'] = date_list_simulation
            df_simulation['municipality'] = municipality
            df_simulation['municipality_id'] = id_municipality
            df_simulation['max_date_to_fit'] = max_date_to_fit
            df_simulation['estimation'] = estimation                
            df_simulations = pd.concat([df_simulations, df_simulation])

municipality

result

len(D_new_deaths)

len(estimated_days_between_infections)

max_date_to_fit

type(estimated_days_between_infections_in_forecast_period)

df_deaths_forecasted

## Saving data

### Forecasted deaths 

df_deaths_forecasted.to_csv('data/df_deaths_forecasted.csv', index=False)

df_forecasting_mean = df_deaths_forecasted.groupby(['municipality','municipality_id','max_date_to_fit','date'])['forecast'].mean().reset_index()
df_forecasting_mean.to_csv('data/df_deaths_forecasted_mean.csv')

### Simulated time series

df_simulations.to_csv('data/df_simulations.csv', index=False)
df_simulations_mean = df_simulations.drop(columns=['estimation']).groupby(['municipality','municipality_id','max_date_to_fit','date']).mean().reset_index()
df_simulations_mean.to_csv('data/df_simulations_mean.csv', index=False)
