==> code/sirds_hybrid/04_extracting_confidence_interval.ipynb <==
# Analyzing the forecast confidence interval

import numpy as np
import pandas as pd
from code.util import calculate_confidence_interval

## Reading data

df_forecasting = pd.read_csv('data/df_deaths_forecasted.csv')
df_forecasting['max_date_to_fit'] = pd.to_datetime(df_forecasting['max_date_to_fit']) 
df_forecasting['date'] = pd.to_datetime(df_forecasting['date'])

## Calculating 

df_forecasting.columns

df_confidence_interval = pd.DataFrame()
for municipality in df_forecasting['municipality'].unique():    
    df_municipality = df_forecasting.loc[df_forecasting['municipality'] == municipality]
    df_municipality = df_municipality.sort_values('date')
    municipality_id = df_municipality['municipality_id'].values[0]
    for date in df_municipality['date'].unique():
        df_municipality_date = df_municipality.loc[df_municipality['date'] == date]
        max_date_to_fit = df_municipality_date['max_date_to_fit'].values[0]
        forecast = df_municipality_date['forecast'].values
        lower_bound, upper_bound = calculate_confidence_interval(forecast)
        mean = np.mean(forecast)
        entry = {'municipality': municipality, 'municipality_id': municipality_id, 'date': date, 'forecast': mean, 'lower_bound': lower_bound, 'upper_bound': upper_bound, 'max_date_to_fit': max_date_to_fit}
        df_confidence_interval = pd.concat([df_confidence_interval, pd.DataFrame.from_records([entry])])

## Saving data

df_confidence_interval.to_csv('data/df_forecast_death_confidence_interval.csv')

df_confidence_interval

df_confidence_interval.columns
